chore: focus on ipinfo-lite only for now
