+++ src/Api/Services/IPInfoLite.php
@@ -15,7 +15,7 @@ use FoF\GeoIP\Api\GeoIP;
 use FoF\GeoIP\Api\ServiceResponse;
 use Psr\Http\Message\ResponseInterface;
 
-class IPInfoLite extends IPInfo
+class IPInfoLite extends BaseGeoService
 {
     protected $host = 'https://api.ipinfo.io';
     protected $settingPrefix = 'fof-geoip.services.ipinfo-lite';
@@ -37,9 +37,17 @@ class IPInfoLite extends IPInfo
         return false;
     }
 
+    protected function buildBatchUrl(array $ips, ?string $apiKey): string
+    {
+        // Not supported by this provider
+        return '';
+    }
+
     protected function getRequestOptions(?string $apiKey, ?array $ips = null): array
     {
-        $options = parent::getRequestOptions($apiKey, $ips);
+        $options = [];
+
+        $options['http_errors'] = false;
         $options['query']['token'] = $apiKey;
 
         return $options;
@@ -59,8 +67,25 @@ class IPInfoLite extends IPInfo
         return $response;
     }
 
+    protected function parseBatchResponse(array $body): array
+    {
+        // Not supported by this provider
+        return [];
+    }
+
     protected function handleError(ResponseInterface $response, object $body): ?ServiceResponse
     {
         return GeoIP::setError('ipinfo-lite', $body->error->message ?? json_encode($body));
+    }
+
+    protected function hasError(ResponseInterface $response, mixed $body): bool
+    {
+        $code = $response->getStatusCode();
+
+        return $code < 200 || $code >= 300 || isset($body?->error);
+    }
+
+    protected function updateRateLimitsFromResponse(ResponseInterface $response, string $requestType = 'single'): void
+    {
     }
 }
